--- swipeswap/lib/screens/user.dart
@@ -1,23 +1,0 @@
-import 'package:flutter/material.dart';
-
-/// User profile page
-/// Users have:
-/// - currentLocation
-/// - dateCreated
-/// - fullName
-/// - phoneNumber
-/// - sellerRating
-/// - TODO: profilePic
-class User extends StatefulWidget {
-  const User({super.key});
-
-  @override
-  State<User> createState() => _UserState();
-}
-
-class _UserState extends State<User> {
-  @override
-  Widget build(BuildContext context) {
-    return const Placeholder();
-  }
-}
